Fix local-loglet params generation (#3335)

+++ crates/bifrost/src/providers/local_loglet/provider.rs
@@ -105,7 +105,7 @@ impl LogletProvider for LocalLogletProvider {
             .map(|c| c.tail_index().next())
             .unwrap_or(SegmentIndex::OLDEST);
         Ok(LogletParams::from(
-            LogletId::new(log_id, new_segment_index).to_string(),
+            u64::from(LogletId::new(log_id, new_segment_index)).to_string(),
         ))
     }
